refactor: adds entities refactor
