+++ src/backoffice/orders/controllers/OrdersController.ts
@@ -1,6 +1,5 @@
 import GetOrdersService from "../services/GetOrdersService"
 
-
 class OrdersController {
   async get(): Promise<any> {
     const getOrdersService = new GetOrdersService()
