+++ src/backoffice/orders/handler.ts
@@ -8,7 +8,6 @@ export const getOrders = async (_request: FastifyRequest, reply: FastifyReply) =
     const response = await ordersController.get()
     reply.code(response.statusCode).send(MessageUtil.success(response.message, response.statusCode, response.data, {}))
   } catch (error) {
-      console.log(error)
     reply
       .code(error.responseStatusCode)
       .send(MessageUtil.error(error.message, error.responseStatusCode, {}, error.errorDetails))
